création PAGE ADMIN (affiche article + modification et suppression, affiche membres) en cours de finition

+++ src/view/post_view.php
@@ -12,10 +12,20 @@
 				<p><?= nl2br($post->getContent()) ?></p>			
 			</div>
 			<br>
+			<?php
+			if($this->session->get('role') === 'admin') {
+			?>
 			<div>
 				<a href="index.php?action=editPost&amp;id=<?= $post->getId(); ?>">Modifier</a>
 				<a href="index.php?action=deletePost&amp;id=<?= $post->getId(); ?>">Supprimer</a>
 			</div>
+			<br>
+			<div>
+				<a href="index.php?action=administration">Retour à l'accueil</a>
+			</div>
+			<?php
+			}
+			?>
 		</div>
 	</div>
 </div>
